Divided index.js into multiple files.
