+++ src/utils/owlterms.js
@@ -1,0 +1,10 @@
+
+// Some OWL constants to be used.
+module.exports = {
+  CDAO_HAS_CHILD: 'obo:CDAO_0000149',
+  CDAO_HAS_DESCENDANT: 'obo:CDAO_0000174',
+  PHYLOREF_EXCLUDES_LINEAGE_TO: 'phyloref:excludes_lineage_to',
+  PHYLOREFERENCE_TEST_CASE: 'testcase:PhyloreferenceTestCase',
+  PHYLOREFERENCE_PHYLOGENY: 'testcase:PhyloreferenceTestPhylogeny',
+  TESTCASE_SPECIFIER: 'testcase:Specifier',
+};
